third_party/blink/web_tests/fast/js/mozilla/strict/8.12.5-expected.txt

PASS 'use strict'; var o = obj(); o.all = 2; o.all is 2
PASS var o = obj(); o.all = 2; o.all is 2
PASS true === true
PASS 'use strict'; var o = obj(); o.nowrite = 2; o.nowrite threw exception of type TypeError.
PASS var o = obj(); o.nowrite = 2; o.nowrite is 1
PASS true === true
PASS 'use strict'; var o = obj(); o.noconfig = 2; o.noconfig is 2
PASS var o = obj(); o.noconfig = 2; o.noconfig is 2
PASS true === true
PASS 'use strict'; var o = obj(); o.noble = 2; o.noble threw exception of type TypeError.
PASS var o = obj(); o.noble = 2; o.noble is 1
PASS true === true
PASS 'use strict'; obj().nowrite++ threw exception of type TypeError.
PASS obj().nowrite++ is 1
PASS true === true
PASS 'use strict'; ++obj().nowrite threw exception of type TypeError.
PASS ++obj().nowrite is 2
PASS true === true
PASS 'use strict'; obj().nowrite-- threw exception of type TypeError.
PASS obj().nowrite-- is 1
PASS true === true
PASS 'use strict'; --obj().nowrite threw exception of type TypeError.
PASS --obj().nowrite is 0
PASS true === true
PASS 'use strict'; var a = arr(); a[0] = 2; a[0] is 2
PASS var a = arr(); a[0] = 2; a[0] is 2
PASS true === true
PASS 'use strict'; var a = arr(); a[1] = 2; a[1] threw exception of type TypeError.
PASS var a = arr(); a[1] = 2; a[1] is 1
PASS true === true
PASS 'use strict'; var a = arr(); a[2] = 2; a[2] is 2
PASS var a = arr(); a[2] = 2; a[2] is 2
PASS true === true
PASS 'use strict'; var a = arr(); a[3] = 2; a[3] threw exception of type TypeError.
PASS var a = arr(); a[3] = 2; a[3] is 1
PASS true === true
PASS 'use strict'; arr()[1]++ threw exception of type TypeError.
PASS arr()[1]++ is 1
PASS true === true
PASS 'use strict'; ++arr()[1] threw exception of type TypeError.
PASS ++arr()[1] is 2
PASS true === true
PASS 'use strict'; arr()[1]-- threw exception of type TypeError.
PASS arr()[1]-- is 1
PASS true === true
PASS 'use strict'; --arr()[1] threw exception of type TypeError.
PASS --arr()[1] is 0
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
